+++ backend/src/main/java/com/team5/secondhand/global/auth/AuthenticationFilter.java
@@ -2,7 +2,7 @@ package com.team5.secondhand.global.auth;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.team5.secondhand.application.member.dto.response.MemberDetails;
-import com.team5.secondhand.global.jwt.util.JwtProperties;
+import com.team5.secondhand.global.properties.JwtProperties;
 import com.team5.secondhand.global.jwt.util.JwtUtils;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
